+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.66.1
 * Added flush messages from buffer before close topic writer
 * Added Flush method for topic writer